initial commit for fdc3 handlers refactor
