+++ packages/main/src/handlers/fdc3/1_2/topics.ts
@@ -1,0 +1,20 @@
+export const FDC3_TOPICS = {
+  DROP_CONTEXT_LISTENER: 'FDC3:dropContextListener',
+  ADD_CONTEXT_LISTENER: 'FDC3:addContextListener',
+  DROP_INTENT_LISTENER: 'FDC3:dropIntentListener',
+  ADD_INTENT_LISTENER: 'FDC3:addIntentListener',
+  BROADCAST: 'FDC3:broadcast',
+  OPEN: 'FDC3:open',
+  CONTEXT: 'FDC3:context',
+  GET_CURRENT_CONTEXT: 'FDC3:getCurrentContext',
+  GET_OR_CREATE_CHANNEL: 'FDC3:getOrCreateChannel',
+  GET_SYSTEM_CHANNELS: 'FDC3:getSystemChannels',
+  LEAVE_CURRENT_CHANNEL: 'FDC3:leaveCurrentChannel',
+  GET_CURRENT_CHANNEL: 'FDC3:getCurrentChannel',
+  JOIN_CHANNEL: 'FDC3:joinChannel',
+  INTENT: 'FDC3:intent',
+  FIND_INTENT: 'FDC3:findIntent',
+  FIND_INTENTS_BY_CONTEXT: 'FDC3:findIntentsByContext',
+  RAISE_INTENT: 'FDC3:raiseIntent',
+  RAISE_INTENTS_FOR_CONTEXT: 'FDC3:raiseIntentsForContext',
+};
