+++ packages/main/tsconfig.json
@@ -14,7 +14,7 @@
 
     "baseUrl": ".",
     "paths": {
-      "/@/*": ["./src/*"]
+      "/@/*": ["./src/*"],
     }
   },
   "include": ["src/**/*.ts", "../../types/**/*.d.ts"],
